refactor engine.py
