+++ src/carapace/session/__init__.py
@@ -2,7 +2,8 @@
 
 from __future__ import annotations
 
-from carapace.session.engine import ActiveSession, SessionEngine, SessionSubscriber
+from carapace.session.engine import SessionEngine
 from carapace.session.manager import SessionManager
+from carapace.session.types import ActiveSession, SessionSubscriber
 
 __all__ = ["ActiveSession", "SessionEngine", "SessionManager", "SessionSubscriber"]
